added data and snakemake subworkflow
