+++ barcode_platemap.csv
@@ -1,0 +1,10 @@
+Assay_Plate_Barcode,Plate_Map_Name
+SQ00014610,DEPENDENCIES1_A549
+SQ00014611,DEPENDENCIES1_A549
+SQ00014612,DEPENDENCIES1_A549
+SQ00014613,DEPENDENCIES1_ES2
+SQ00014614,DEPENDENCIES1_ES2
+SQ00014615,DEPENDENCIES1_ES2
+SQ00014616,DEPENDENCIES1_HCC44
+SQ00014617,DEPENDENCIES1_HCC44
+SQ00014618,DEPENDENCIES1_HCC44
