+++ data/metadata/platemap/DEPENDENCIES1_raw.csv
@@ -1,0 +1,1 @@
+Row,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24A,EMPTY,MCL1-5,AKT1-1,KRAS-2B,AKT1-2,EMPTY,BRAF1-1,EMPTY,BRAF-2,MYC-2,BRD4-1,MYC-1,BRD4-2,MCL1-2,CDK4-1,BCL2L2-4,CDK4-2,MAPK1-2,CHEK2-1,MAPK1-1,CHEK2-2,KRAS-2A,CSNK1A1-1,KRAS-1B,Chr2-1,PPIB-2,POLR2D-2,EMPTY,Chr2-4,PSMA1-2,PSMA1-1,Chr2-5,EMPTY,ORC4-1,EMPTY,Chr2-2,Luc-1,EMPTY,Luc-1,EMPTY,EMPTY,EMPTY,EMPTY,EMPTY,PPIB-1,LacZ-3,EMPTY,LacZ-3C,CSNK1A1-2,HIF1a-2,CTNNB1-1,HIF1a-1,CTNNB1-2,GPX4-2,DOT1L-1,GPX4-1,DOT1L-2,GLS-2,EGFR-1,GLS-1,EGFR-2,EMPTY,ERBB3-1,EMPTY,ERBB3-2,ERBB2-2,ITGAV-1,ERBB2-1,ITGAV-2,EGLN1-9,JAK3-1,EGLN1-8D,Chr2-1,PPIB-2,POLR2D-2,EMPTY,Chr2-4,PSMA1-2,PSMA1-1,Chr2-5,PRDX1-1,ORC4-1,PRDX1-2,Chr2-2,Luc-1,YAP1-2,Luc-2,YAP1-1,EMPTY,EMPTY,EMPTY,EMPTY,PPIB-1,LacZ-2,EMPTY,LacZ-2E,JAK3-2,COX5B-2,KIF11-1,COX5B-1,KIF11-2,CDK2-2,MTOR-1,CDK2-1,MTOR-2,CCNE1-2,PIK3CA-1,CCNE1-1,PIK3CA-2,CCND1-2,RAC1-1,CCND1-1,RAC1-2,BRCA1-2,RHOA-1,BRCA1-1,RHOA-2,BCL2L2-2,SDHB-1,BCL2L2-1F,POLR2D-1,EMPTY,Chr2-3,PPIB-2,ORC4-2,Chr2-6,Chr2-6,ORC4-2,RAF1-1,Chr2-3,RAF1-2,POLR2D-1,Chr2-1,XBP-2,POLR2D-2,XBP-1,Chr2-4,PSMA1-2,PSMA1-1,Chr2-5,EMPTY,ORC4-1,PPIB-1,Chr2-2G,SDHB-2,BCL2-2,SMARCB1-1,BCL2-1,SMARCB1-2,AURKB-2,STAT3-1,CDK4-3,STAT3-2,ATP50-2,SUZ12-1,ATP50-1,SUZ12-2,ATF4-2,EZH2-1,ATF4-1,EZH2-2,ARID1B-2,PTK2-1,ARID1B-1,PTK2-2,EMPTY,EMPTY,EMPTYH,POLR2D-1,EMPTY,Chr2-3,PPIB-2,ORC4-2,Chr2-6,Chr2-6,ORC4-2,SLC2A1-1,Chr2-3,SLC2A2-2,POLR2D-1,Chr2-1,TXN-2,POLR2D-2,TXN-1,Chr2-4,PSMA1-2,PSMA1-1,Chr2-5,EMPTY,ORC4-1,PPIB-1,Chr2-2I,EMPTY,EMPTY,EMPTY,PTK2-2,ARID1B-1,PTK2-1,ARID1B-2,EZH2-2,ATF4-1,EZH2-1,ATF4-2,SUZ12-2,ATP50-1,SUZ12-1,ATP50-2,STAT3-2,CDK4-3,STAT3-1,AURKB-2,SMARCB1-2,BCL2-1,SMARCB1-1,BCL2-2,SDHB-2J,Chr2-2,PPIB-1,ORC4-1,EMPTY,Chr2-5,PSMA1-1,PSMA1-2,Chr2-4,TXN-1,POLR2D-2,TXN-2,Chr2-1,POLR2D-1,SLC2A2-2,Chr2-3,SLC2A1-1,ORC4-2,Chr2-6,Chr2-6,ORC4-2,PPIB-2,Chr2-3,EMPTY,POLR2D-1K,BCL2L2-1,SDHB-1,BCL2L2-2,RHOA-2,BRCA1-1,RHOA-1,BRCA1-2,RAC1-2,CCND1-1,RAC1-1,CCND1-2,PIK3CA-2,CCNE1-1,PIK3CA-1,CCNE1-2,MTOR-2,CDK2-1,MTOR-1,CDK2-2,KIF11-2,COX5B-1,KIF11-1,COX5B-2,JAK3-2L,Chr2-2,PPIB-1,ORC4-1,EMPTY,Chr2-5,PSMA1-1,PSMA1-2,Chr2-4,XBP-1,POLR2D-2,XBP-2,Chr2-1,POLR2D-1,RAF1-2,Chr2-3,RAF1-1,ORC4-2,Chr2-6,Chr2-6,ORC4-2,PPIB-2,Chr2-3,EMPTY,POLR2D-1M,EGLN1-8,JAK3-1,EGLN1-9,ITGAV-2,ERBB2-1,ITGAV-1,ERBB2-2,ERBB3-2,EMPTY,ERBB3-1,EMPTY,EGFR-2,GLS-1,EGFR-1,GLS-2,DOT1L-2,GPX4-1,DOT1L-1,GPX4-2,CTNNB1-2,HIF1a-1,CTNNB1-1,HIF1a-2,CSNK1A1-2N,LacZ-2,EMPTY,LacZ-2,PPIB-1,EMPTY,EMPTY,EMPTY,EMPTY,YAP1-1,Luc-2,YAP1-2,Luc-1,Chr2-2,PRDX1-2,ORC4-1,PRDX1-1,Chr2-5,PSMA1-1,PSMA1-2,Chr2-4,EMPTY,POLR2D-2,PPIB-2,Chr2-1O,KRAS-1,CSNK1A1-1,KRAS-2A,CHEK2-2,MAPK1-1,CHEK2-1,MAPK1-2,CDK4-2,BCL2L2-4,CDK4-1,MCL1-2,BRD4-2,MYC-1,BRD4-1,MYC-2,BRAF-2,EMPTY,BRAF1-1,EMPTY,AKT1-2,KRAS-2B,AKT1-1,MCL1-5,EMPTYP,LacZ-3,EMPTY,LacZ-3,PPIB-1,EMPTY,EMPTY,EMPTY,EMPTY,EMPTY,Luc-2,EMPTY,Luc-1,Chr2-2,EMPTY,ORC4-1,EMPTY,Chr2-5,PSMA1-1,PSMA1-2,Chr2-4,EMPTY,POLR2D-2,PPIB-2,Chr2-1
