+++ data/metadata/platemap/broad_sample_map.csv
@@ -1,0 +1,109 @@
+well_position,pert_name,broad_sample
+C01,CSNK1A1-2,BRDN0000585480
+E01,JAK3-2,BRDN0001065702
+G01,SDHB-2,BRDN0000986593
+N01,LacZ-2,BRDN0000585793
+P01,LacZ-3,BRDN0000585800
+K02,SDHB-1,BRDN0000986589
+M02,JAK3-1,BRDN0001065886
+O02,CSNK1A1-1,BRDN0001065659
+A03,AKT1-1,BRDN0001054908
+C03,CTNNB1-1,BRDN0001065286
+E03,KIF11-1,BRDN0000734411
+G03,SMARCB1-1,BRDN0001062203
+N03,LacZ-2,BRDN0000585793
+P03,LacZ-3,BRDN0000585800
+I04,PTK2-2,BRDN0000990405
+K04,RHOA-2,BRDN0000990437
+M04,ITGAV-2,BRDN0001052667
+O04,CHEK2-2,BRDN0000585854
+A05,AKT1-2,BRDN0001055115
+C05,CTNNB1-2,BRDN0000579714
+E05,KIF11-2,BRDN0000734657
+G05,SMARCB1-2,BRDN0001062218
+I06,PTK2-1,BRDN0000990543
+K06,RHOA-1,BRDN0000990371
+M06,ITGAV-1,BRDN0001052092
+O06,CHEK2-1,BRDN0000586287
+A07,BRAF1-1,BRDN0000986682
+C07,DOT1L-1,BRDN0001057300
+E07,MTOR-1,BRDN0000734637
+G07,STAT3-1,BRDN0001063739
+I08,EZH2-2,BRDN0000733144
+K08,RAC1-2,BRDN0001062199
+M08,ERBB3-2,BRDN0001054867
+O08,CDK4-2,BRDN0001053895
+A09,BRAF-2,BRDN0000986702
+C09,DOT1L-2,BRDN0001057286
+E09,MTOR-2,BRDN0000734053
+G09,STAT3-2,BRDN0000579974
+I10,EZH2-1,BRDN0000733051
+K10,RAC1-1,BRDN0001062183
+M10,ERBB3-1,BRDN0001054818
+N10,Luc-2,BRDN0000585515
+O10,CDK4-1,BRDN0001053869
+P10,Luc-2,BRDN0000585515
+A11,BRD4-1,BRDN0000733514
+C11,EGFR-1,BRDN0001054960
+E11,PIK3CA-1,BRDN0001054759
+G11,SUZ12-1,BRDN0001057351
+I12,SUZ12-2,BRDN0001057661
+K12,PIK3CA-2,BRDN0001054804
+M12,EGFR-2,BRDN0001054761
+N12,Luc-1,BRDN0000585635
+O12,BRD4-2,BRDN0000733577
+P12,Luc-1,BRDN0000585635
+A13,BRD4-2,BRDN0000733577
+B13,Luc-1,BRDN0000585635
+C13,EGFR-2,BRDN0001054761
+D13,Luc-1,BRDN0000585635
+E13,PIK3CA-2,BRDN0001054804
+G13,SUZ12-2,BRDN0001057661
+I14,SUZ12-1,BRDN0001057351
+K14,PIK3CA-1,BRDN0001054759
+M14,EGFR-1,BRDN0001054960
+O14,BRD4-1,BRDN0000733514
+A15,CDK4-1,BRDN0001053869
+B15,Luc-1,BRDN0000585635
+C15,ERBB3-1,BRDN0001054818
+D15,Luc-2,BRDN0000585515
+E15,RAC1-1,BRDN0001062183
+G15,EZH2-1,BRDN0000733051
+I16,STAT3-2,BRDN0000579974
+K16,MTOR-2,BRDN0000734053
+M16,DOT1L-2,BRDN0001057286
+O16,BRAF-2,BRDN0000986702
+A17,CDK4-2,BRDN0001053895
+C17,ERBB3-2,BRDN0001054867
+E17,RAC1-2,BRDN0001062199
+G17,EZH2-2,BRDN0000733144
+I18,STAT3-1,BRDN0001063739
+K18,MTOR-1,BRDN0000734637
+M18,DOT1L-1,BRDN0001057300
+O18,BRAF1-1,BRDN0000986682
+A19,CHEK2-1,BRDN0000586287
+C19,ITGAV-1,BRDN0001052092
+E19,RHOA-1,BRDN0000990371
+G19,PTK2-1,BRDN0000990543
+I20,SMARCB1-2,BRDN0001062218
+K20,KIF11-2,BRDN0000734657
+M20,CTNNB1-2,BRDN0000579714
+O20,AKT1-2,BRDN0001055115
+A21,CHEK2-2,BRDN0000585854
+C21,ITGAV-2,BRDN0001052667
+E21,RHOA-2,BRDN0000990437
+G21,PTK2-2,BRDN0000990405
+B22,LacZ-3,BRDN0000585800
+D22,LacZ-2,BRDN0000585793
+I22,SMARCB1-1,BRDN0001062203
+K22,KIF11-1,BRDN0000734411
+M22,CTNNB1-1,BRDN0001065286
+O22,AKT1-1,BRDN0001054908
+A23,CSNK1A1-1,BRDN0001065659
+C23,JAK3-1,BRDN0001065886
+E23,SDHB-1,BRDN0000986589
+B24,LacZ-3,BRDN0000585800
+D24,LacZ-2,BRDN0000585793
+I24,SDHB-2,BRDN0000986593
+K24,JAK3-2,BRDN0001065702
+M24,CSNK1A1-2,BRDN0000585480
