+++ envs/cytominer_env.yaml
@@ -1,0 +1,10 @@
+name: ctyominer_env
+channels:
+  - conda-forge
+  - bioconda
+  - defaults
+dependencies:
+  - python=3.7.0
+  - pip
+  - pip:
+    - "git+https://github.com/cytomining/pycytominer@1279cd2349715604d226b58e7aa61a8a8eda7e54"
